+++ carePLUS/src/main/java/com/kosta/care/entity/Adminssion.java
@@ -21,15 +21,15 @@ import lombok.NoArgsConstructor;
 public class Adminssion {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer admNum;
+	private Long admNum;
 	@Column
-	private Integer docNum;
+	private Long docNum;
 	@Column
-	private Integer bedsId;
+	private Long bedsId;
 	@Column
-	private Integer patNum;
+	private Long patNum;
 	@Column
-	private Integer jobNum;
+	private Long jobNum;
 	@Column
 	private Date adminssionDate;
 	@Column
